v2/source codes/phasesegmentation/AFR_Project_2/GJ/requirements.txt
absl-py==1.1.0
av==9.2.0
black==22.3.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cycler==0.11.0
docker-pycreds==0.4.0
fonttools==4.33.3
fvcore==0.1.5.post20220512
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
idna==3.3
importlib-metadata==4.12.0
iopath==0.1.9
joblib==1.1.0
kiwisolver==1.4.3
Markdown==3.3.7
matplotlib==3.5.2
mypy-extensions==0.4.3
networkx==2.8.4
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
parameterized==0.8.1
pathspec==0.9.0
pathtools==0.1.2
Pillow==9.1.1
platformdirs==2.5.2
portalocker==2.4.0
promise==2.3
protobuf==3.19.4
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytorchvideo==0.1.5
pytz==2022.1
pywin32==304
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
sentry-sdk==1.6.0
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
tabulate==0.8.10
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.1.0
timm==0.5.4
tomli==2.0.1
torch-tb-profiler==0.4.0
tqdm==4.64.0
typing-extensions==4.2.0
urllib3==1.26.9
wandb==0.12.20
Werkzeug==2.1.2
yacs==0.1.8
zipp==3.8.0
